Sub2: changed int to long for larger k and a[i]

+++ ABC353/c.cpp
@@ -12,13 +12,15 @@ int main() {
 		cin >> a[i];
 	}
 
-	int sum = 0;
+	// int: 10^9.3 at most; long: 10^18.9 at most
+	// if n = 3*10^5, each f is 1e8-1 at most, so 3e5 * (1e8-1) = 10^13.5, overflow in int 
+	long sum = 0;
 	for (int i = 1; i < n; i++) {
 		// int i_sum = 0
 		for (int j = i+1; j < n+1; j++) {
 			// i, j is 1-indexed above, but A is 0-indexed.
 			// A_1 (1-indexed) = A_0 (0-indexed)
-			sum += (a[i-1] + a[j-1]) % 100000000;
+			sum += (long)(a[i-1] + a[j-1]) % 100000000; // a[i-1], a[j-1] is 10^8 at mose; int is 2^31 = 10^9.3, so no overflow
 		}
 	}
